Move existing field to new structure #j9rt2
